break out hql keywords
